Move enum class VendorId out of OnboardingPayload
