+++ examples/android/CHIPTool/app/src/main/java/com/google/chip/chiptool/setuppayloadscanner/QrCodeInfo.kt
@@ -1,7 +1,7 @@
 package com.google.chip.chiptool.setuppayloadscanner
 
 import android.os.Parcelable
-import chip.onboardingpayload.OptionalQRCodeInfo.OptionalQRCodeInfoType
+import chip.onboardingpayload.OptionalQRCodeInfoType
 import kotlinx.parcelize.Parcelize
 
 @Parcelize
