+++ src/controller/java/src/chip/onboardingpayload/OptionalQRCodeInfo.kt
@@ -17,18 +17,18 @@
  
 package chip.onboardingpayload
 
+enum class OptionalQRCodeInfoType {
+  TYPE_UNKNOWN,
+  TYPE_STRING,
+  TYPE_INT32,
+  TYPE_INT64,
+  TYPE_UINT32,
+  TYPE_UINT64
+}
+
 data class OptionalQRCodeInfo(
   var tag: Int = 0,
   var type: OptionalQRCodeInfoType = OptionalQRCodeInfoType.TYPE_UNKNOWN,
   var data: String? = null,
   var int32: Int = 0
-) {
-  enum class OptionalQRCodeInfoType {
-    TYPE_UNKNOWN,
-    TYPE_STRING,
-    TYPE_INT32,
-    TYPE_INT64,
-    TYPE_UINT32,
-    TYPE_UINT64
-  }
-}
+)
